2015: d19p2

+++ 2015/src/day19/day19.h
@@ -4,4 +4,4 @@
 
 void day19_main(void);
 uint16_t day19_part1(const char* const*);
-uint16_t day19_part2(const char* const*);
+uint8_t day19_part2(const char* const*);
